ScraperService/requirements.txt
requests
beautifulsoup4
feedparser
kafka-python
schedule
python-dotenv
redis
python-dateutil
telethon
psycopg2